AppAutomateDemo/requirements.txt
Appium-Python-Client==2.10.1
pytest==7.3.1
pytest-xdist==3.3.1
python-dotenv==1.0.0
selenium==4.9.1